+++ ml-agents/mlagents/trainers/tests/test_learn.py
@@ -1,4 +1,5 @@
 import pytest
+import yaml
 from unittest.mock import MagicMock, patch, mock_open
 from mlagents.trainers import learn
 from mlagents.trainers.trainer_controller import TrainerController
@@ -13,6 +14,25 @@ def basic_options(extra_args=None):
     if extra_args:
         args += [f"{k}={v}" for k, v in extra_args.items()]
     return parse_command_line(args)
+
+
+MOCK_YAML = """
+    behaviors:
+        {}
+    """
+
+MOCK_SAMPLER_CURRICULUM_YAML = """
+    behaviors:
+        behavior1:
+            curriculum:
+                curriculum1
+        behavior2:
+            curriculum:
+                curriculum2
+
+    parameter_randomization:
+        sampler1
+    """
 
 
 @patch("mlagents.trainers.learn.write_timing_tree")
@@ -37,8 +57,7 @@ def test_run_training(
     mock_env.external_brain_names = []
     mock_env.academy_name = "TestAcademyName"
     create_environment_factory.return_value = mock_env
-    trainer_config_mock = MagicMock()
-    load_config.return_value = trainer_config_mock
+    load_config.return_value = yaml.safe_load(MOCK_YAML)
 
     mock_init = MagicMock(return_value=None)
     with patch.object(TrainerController, "__init__", mock_init):
@@ -74,7 +93,7 @@ def test_bad_env_path():
         )
 
 
-@patch("builtins.open", new_callable=mock_open, read_data="{}")
+@patch("builtins.open", new_callable=mock_open, read_data=MOCK_YAML)
 def test_commandline_args(mock_file):
 
     # No args raises
@@ -83,10 +102,9 @@ def test_commandline_args(mock_file):
 
     # Test with defaults
     opt = parse_command_line(["mytrainerpath"])
-    assert opt.trainer_config == {}
+    assert opt.behaviors == {}
     assert opt.env_path is None
-    assert opt.curriculum_config is None
-    assert opt.sampler_config is None
+    assert opt.parameter_randomization is None
     assert opt.keep_checkpoints == 5
     assert opt.lesson == 0
     assert opt.resume is False
@@ -103,8 +121,6 @@ def test_commandline_args(mock_file):
     full_args = [
         "mytrainerpath",
         "--env=./myenvfile",
-        "--curriculum=./mycurriculum",
-        "--sampler=./mysample",
         "--keep-checkpoints=42",
         "--lesson=3",
         "--resume",
@@ -120,10 +136,9 @@ def test_commandline_args(mock_file):
     ]
 
     opt = parse_command_line(full_args)
-    assert opt.trainer_config == {}
+    assert opt.behaviors == {}
     assert opt.env_path == "./myenvfile"
-    assert opt.curriculum_config == {}
-    assert opt.sampler_config == {}
+    assert opt.parameter_randomization is None
     assert opt.keep_checkpoints == 42
     assert opt.lesson == 3
     assert opt.run_id == "myawesomerun"
@@ -137,7 +152,13 @@ def test_commandline_args(mock_file):
     assert opt.resume is True
 
 
-@patch("builtins.open", new_callable=mock_open, read_data="{}")
+@patch("builtins.open", new_callable=mock_open, read_data=MOCK_SAMPLER_CURRICULUM_YAML)
+def test_sampler_configs(mock_file):
+    opt = parse_command_line(["mytrainerpath"])
+    assert opt.parameter_randomization == "sampler1"
+
+
+@patch("builtins.open", new_callable=mock_open, read_data=MOCK_YAML)
 def test_env_args(mock_file):
     full_args = [
         "mytrainerpath",